Move preview generator tests to a dedicated file
